refactor(qa): 検索結果の構造を改善し、インデックス管理を追加
